Se modifica la forma de verificar el jwt y permisos

+++ src/graphql/user-role/userRole.resolver.js
@@ -1,41 +1,36 @@
 import checkRoles from '../../utils/auth/checkRolesGql.js';
 import { checkJwtGql } from '../../utils/auth/checkJwtGql.js';
 import RoleService from '../../services/userRole.services.js';
+import { validateUserIsAdmin } from '../../utils/auth/userPermissions.js';
 
 const service = new RoleService();
 
 export const addRole = async (_, { roleName }, context) => {
-  const user = await checkJwtGql(context);
-  checkRoles(user, 'admin');
+  await validateUserIsAdmin(context);
   return service.create(roleName);
 };
 
 export const allRoles = async (_, __, context) => {
-  const user = await checkJwtGql(context);
-  checkRoles(user, 'admin');
+  await validateUserIsAdmin(context);
   return service.find();
 };
 
 export const roleById = async (_, { idRole }, context) => {
-  const user = await checkJwtGql(context);
-  checkRoles(user, 'admin', 'customer');
+  await checkJwtGql(context);
   return service.findOneById(idRole);
 };
 
 export const roleByName = async (_, { roleName }, context) => {
-  const user = await checkJwtGql(context);
-  checkRoles(user, 'admin', 'customer');
+  await validateUserIsAdmin(context);
   return service.findByName(roleName);
 };
 
 export const updateRole = async (_, args, context) => {
-  const user = await checkJwtGql(context);
-  checkRoles(user, 'admin');
+  await validateUserIsAdmin(context);
   return service.update(args);
 };
 
 export const deleteRole = async (_, { idRole }, context) => {
-  const user = await checkJwtGql(context);
-  checkRoles(user, 'admin');
+  await validateUserIsAdmin(context);
   return service.delete(idRole);
 };
